requirements.txt
flask
flask-sqlalchemy
requests
google-genai     
python-dotenv
apscheduler
psycopg2-binary #neon db
pytz #for conversion to EST